Refactoring : move ActionWrapper class to single file
