+++ src/BlazorQuery.Library/ActionWrapper.cs
@@ -1,0 +1,21 @@
+﻿using Microsoft.JSInterop;
+using System;
+
+namespace BlazorQuery.Library
+{
+    public class ActionWrapper<T>
+    {
+        private readonly Action<T> _action;
+
+        public ActionWrapper(Action<T> action)
+        {
+            this._action = action;
+        }
+        // Callbacks
+        [JSInvokable]
+        public void ActionCallback(T obj)
+        {
+            _action(obj);
+        }
+    }
+}
