+++ src/BlazorQuery.Library/BlazorQueryDOM.cs
@@ -8,23 +8,6 @@ using System.Threading.Tasks;
 namespace BlazorQuery.Library
 {
 
-    public class ActionWrapper<T>
-    {
-        private readonly Action<T> _action;
-
-        public ActionWrapper(Action<T> action)
-        {
-            this._action = action;
-        }
-        // Callbacks
-        [JSInvokable]
-        public void ActionCallback(T obj)
-        {
-            _action(obj);
-        }
-
-    }
-
     public class BlazorQueryDOM : ComponentBase
     {
         private IJSRuntime JSRuntime { get; set; }
